+++ src/generator/readable.rs
@@ -416,9 +416,8 @@ impl LuaGenerator for ReadableLuaGenerator {
 
         self.raw_push_str(" = ");
 
-        let values = assign.get_values();
-        let last_value_index = values.len() - 1;
-        values.iter().enumerate().for_each(|(index, value)| {
+        let last_value_index = assign.values_len() - 1;
+        assign.iter_values().enumerate().for_each(|(index, value)| {
             self.write_expression(value);
 
             if index != last_value_index {
@@ -447,14 +446,12 @@ impl LuaGenerator for ReadableLuaGenerator {
             }
         });
 
-        let values = assign.get_values();
-
-        if !values.is_empty() {
+        if assign.has_values() {
             self.raw_push_str(" = ");
 
-            let last_value_index = values.len() - 1;
+            let last_value_index = assign.values_len() - 1;
 
-            values.iter().enumerate().for_each(|(index, value)| {
+            assign.iter_values().enumerate().for_each(|(index, value)| {
                 self.write_expression(value);
 
                 if index != last_value_index {